Add files via upload
Prediction using Decision Tree  Algorithm :
Here the aim was to create the Decision Tree classifier and visualize it graphically. 
The purpose is if we feed any new data to this classifier, it would be able to  predict the right class accordingly. 
I have applied decision tree classifier on the Iris dataset.